validity-test-ospath: Different modules for different platforms
