+++ validity-test-ospath/OsPath/Gen/Posix.hs
@@ -1,0 +1,8 @@
+{-# LANGUAGE CPP #-}
+
+#define PLATFORM_NAME        Posix
+-- #define PLATFORM_NAME_STRING "Posix"
+#define PLATFORM_PATH        PosixPath
+-- #define PLATFORM_STRING      PosixString
+-- #define IS_WINDOWS           0
+#include "Include.hs"
