+++ validity-test-ospath/OsPath/Gen/Windows.hs
@@ -1,0 +1,8 @@
+{-# LANGUAGE CPP #-}
+
+#define PLATFORM_NAME        Windows
+-- #define PLATFORM_NAME_STRING "Windows"
+#define PLATFORM_PATH        WindowsPath
+-- #define PLATFORM_STRING      WindowsString
+-- #define IS_WINDOWS           1
+#include "Include.hs"
