+++ validity-test-ospath/Posix.hs
@@ -1,0 +1,5 @@
+{-# LANGUAGE CPP #-}
+
+#define PLATFORM_NAME        Posix
+#define PLATFORM_NAME_STRING "Posix"
+#include "Include.hs"
