+++ validity-test-ospath/Windows.hs
@@ -1,0 +1,5 @@
+{-# LANGUAGE CPP #-}
+
+#define PLATFORM_NAME        Windows
+#define PLATFORM_NAME_STRING "Windows"
+#include "Include.hs"
